Fixed resetting feature

+++ NetworkController/DataTransferStructures/Other/AckStatus.cs
@@ -4,7 +4,7 @@ using System.Text;
 
 namespace NetworkController.DataTransferStructures.Other
 {
-    public enum AckStatus : ushort
+    public enum AckStatus : int
     {
         Success = 0,
         Failure = 1
